+++ packages/pwa-kit-react-sdk/src/ssr/server/react-rendering.js
@@ -150,6 +150,17 @@ export const render = async (req, res, next) => {
         })
     }
 
+    // Some application extensions need to be serialized because they have asynchronous state
+    const serializedExtensions = Object.fromEntries(
+        applicationExtensions.flatMap((extension) => {
+            if (typeof extension.getRoutesAsync !== 'function') return []
+            const routes = extension.serializeAsyncRoutes()
+            // TODO W-18257236: Use a unique key for each extension like the extension ID from
+            // extension-meta.json
+            return [[extension.getName(), {routes}]]
+        })
+    )
+
     // Step 1 - Find the match.
 
     // Call `beforeRouteMatch` application extension hook.
@@ -228,7 +239,8 @@ export const render = async (req, res, next) => {
             res,
             location,
             config,
-            appJSX
+            appJSX,
+            serializedExtensions
         })
     } catch (e) {
         // This is an unrecoverable error.
@@ -293,7 +305,7 @@ const renderToString = (jsx, extractor) =>
     ReactDOMServer.renderToString(extractor.collectChunks(jsx))
 
 const renderApp = (args) => {
-    const {req, res, appStateError, appJSX, appState, config} = args
+    const {req, res, appStateError, appJSX, appState, config, serializedExtensions} = args
     const extractor = new ChunkExtractor({statsFile: BUNDLES_PATH, publicPath: getAssetUrl()})
 
     const ssrOnly = 'mobify_server_only' in req.query || '__server_only' in req.query
@@ -357,6 +369,7 @@ const renderApp = (args) => {
         __CONFIG__: config,
         __PRELOADED_STATE__: appState,
         __ERROR__: error,
+        __EXTENSIONS__: serializedExtensions,
         // `window.Progressive` has a long history at Mobify and some
         // client-side code depends on it. Maintain its name out of tradition.
         Progressive: getWindowProgressive(req, res)