+++ packages/pwa-kit-react-sdk/src/ssr/universal/components/route-component/index.js
@@ -409,7 +409,7 @@ export const getAllRoutes = async (locals = {}) => {
     const extensionRoutes = (
         await Promise.all(
             applicationExtensions.map((extension) =>
-                extension.getRoutesAsync
+                typeof extension.getRoutesAsync === 'function'
                     ? extension.getRoutesAsync({locals})
                     : extension.getRoutes({locals})
             )