+++ packages/pwa-kit-react-sdk/src/ssr/universal/hooks/index.js
@@ -7,7 +7,7 @@
 /* istanbul ignore file */
 
 import React, {useContext} from 'react'
-import {CorrelationIdContext, ServerContext} from '../contexts'
+import {CorrelationIdContext, ServerContext, RoutesContext} from '../contexts'
 
 /**
  * Use this hook to get the correlation id value of the closest CorrelationIdProvider component.
@@ -69,4 +69,17 @@ export const useOrigin = ({fromXForwardedHeader = false}) => {
         return xForwardedOrigin
     }
     return APP_ORIGIN
+}
+
+/**
+ * Use this hook to get the routes value of the closest RoutesProvider component.
+ *
+ * @returns {object} array of routes
+ */
+export const useRoutes = () => {
+    const context = useContext(RoutesContext)
+    if (!context) {
+        throw new Error('useRoutes must be used within a RoutesProvider')
+    }
+    return context
 }