+++ packages/pwa-kit-react-sdk/src/ssr/universal/utils.ts
@@ -18,7 +18,7 @@ import {bundleBasePath} from '@salesforce/pwa-kit-runtime/utils/ssr-namespace-pa
 
 const onClient = typeof window !== 'undefined'
 
-const EXTENIONS_NAMESPACE = '__extensions'
+const EXTENSIONS_NAMESPACE = '__extensions'
 const STATIC_FOLDER = 'static'
 
 type GetAssetUrlOptions = {
@@ -34,12 +34,15 @@ type GetAssetUrlOptions = {
  */
 export const getAssetUrl = (path: string) => {
     /* istanbul ignore next */
-    const publicPath = onClient
+    let publicPath = onClient
         ? // @ts-ignore
           `${window.Progressive.buildOrigin as string}`
-        : `${bundleBasePath}/${process.env.BUNDLE_ID || 'development'}/`
+        : `${bundleBasePath}/${process.env.BUNDLE_ID || 'development'}`
 
-    return path ? `${publicPath}${path}` : publicPath
+    // Normalize the public path by removing the trailing slash
+    publicPath = publicPath.replace(/\/$/, '')
+
+    return path ? `${publicPath}/${path}` : publicPath
 }
 
 // TODO: Once we establish that we have a new @salesforce/pwa-kit-extensibility package, we can move this utility to
@@ -53,23 +56,31 @@ export const getAssetUrl = (path: string) => {
  * @function
  * @returns {string} The full URL to the static asset.
  */
-export const getStaticAssetUrl = (path: string, opts: GetAssetUrlOptions) => {
+export const getStaticAssetUrl = (path = '', opts: GetAssetUrlOptions) => {
     const {appExtensionPackageName = ''} = opts || {}
+    const extensionPrefix = `${
+        appExtensionPackageName ? `/${EXTENSIONS_NAMESPACE}/${appExtensionPackageName}` : ''
+    }`
 
     /* istanbul ignore next */
-    const publicPath = onClient
+    let publicPath = onClient
         ? // @ts-ignore
           `${window.Progressive.buildOrigin as string}`
-        : `${bundleBasePath}/${process.env.BUNDLE_ID || 'development'}/`
+        : `${bundleBasePath}/${process.env.BUNDLE_ID || 'development'}`
 
-    // Ensure all defined path arguments start with `/`.
-    if (path && !path.startsWith('/')) {
-        path = `/${path}`
-    }
+    // Normalize the public path by removing the trailing slash
+    publicPath = publicPath.replace(/\/$/, '')
 
-    return `${publicPath}/${STATIC_FOLDER}${
-        appExtensionPackageName ? `/${EXTENIONS_NAMESPACE}/${appExtensionPackageName}` : ''
-    }${path ? path : ''}`
+    // Ensure path starts with `/`.
+    path = `${path && !path.startsWith('/') ? `/${path}` : path}`
+
+    // Prepend the extension namespace and package if applicable.
+    path = `${extensionPrefix ? `${extensionPrefix}${path}` : path}`
+
+    // Prefix public assets path and static folder.
+    path = `${publicPath}/${STATIC_FOLDER}${path}`
+
+    return path
 }
 
 /**